Add: MainPageContainer component for Wiki/MainPage
